requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cssselect2==0.7.0
distro==1.8.0
fonttools==4.42.1
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
numpy==1.25.2
pandas==2.1.0
Pillow==10.0.0
pycparser==2.21
pydyf==0.7.0
pyee==8.2.2
pyphen==0.14.0
pyppeteer==1.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tabula-py==2.7.0
tinycss2==1.2.1
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
weasyprint==59.0
webencodings==0.5.1
websockets==10.4
zipp==3.17.0
zopfli==0.2.2
